Morgan2

+++ my-app/src/utils/crud/Query.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const getCheckout = gql`
-query Checkout($donationAmount: Int) {
+query Checkout($donationAmount: Float) {
   checkout(donationAmount: $donationAmount) {
     session
   }
